requirements.txt
numpy
pandas
scikit-learn
requests
gitpython
mkdir_p
flask
flask_restful